Fix migration using column existence checks

+++ alembic/versions/24417ff15934_add_created_at_to_settlements.py
@@ -10,6 +10,7 @@ from typing import Sequence, Union
 from alembic import op
 import sqlalchemy as sa
 from sqlalchemy.dialects import postgresql
+from sqlalchemy import inspect
 
 # revision identifiers, used by Alembic.
 revision: str = '24417ff15934'
@@ -17,6 +18,11 @@ down_revision: Union[str, Sequence[str], None] = '001_restructure_add_fields'
 branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None
 
+def column_exists(table_name, column_name):
+    bind = op.get_bind()
+    inspector = inspect(bind)
+    columns = [c["name"] for c in inspector.get_columns(table_name)]
+    return column_name in columns
 
 def upgrade() -> None:
     """Upgrade schema."""
@@ -34,10 +40,8 @@ def upgrade() -> None:
                nullable=False)
     op.create_index(op.f('ix_group_expenses_group_id'), 'group_expenses', ['group_id'], unique=False)
     op.create_index(op.f('ix_group_expenses_paid_by_id'), 'group_expenses', ['paid_by_id'], unique=False)
-    try:
+    if column_exists('group_expenses', 'paid_by_username'):
         op.drop_column('group_expenses', 'paid_by_username')
-    except:
-        pass
     try:
         op.alter_column(
             'group_members',
@@ -64,10 +68,8 @@ def upgrade() -> None:
     op.drop_column('settlements', 'timestamp')
     op.create_index(op.f('ix_splits_group_expense_id'), 'splits', ['group_expense_id'], unique=False)
     op.create_index(op.f('ix_splits_user_id'), 'splits', ['user_id'], unique=False)
-    try:
+    if column_exists('splits', 'paid_by_username'):
         op.drop_column('splits', 'paid_by_username')
-    except:
-        pass
     op.alter_column('users', 'email',
                existing_type=sa.VARCHAR(length=255),
                type_=sa.String(length=100),
